pyproject.toml
[project]
name = "peresteparse"
version = "2.0.0"
description = "Voice transcription to structured flashcards with LLM parsing"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.0",
    "flask-cors==4.0.0",
    "llama-cpp-python>=0.3.0",
    "huggingface-hub>=0.20.0",
    "parakeet-mlx>=0.5.0",
    "numba>=0.60",
    "llvmlite>=0.43",
    "sounddevice>=0.5.0",
    "soundfile>=0.12.0",
    "pywebview==5.3",
    "anthropic>=0.40.0",
    "genanki>=0.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["server"]
omit = ["tests/*"]
